+++ internal/provider/product_type_resource.go
@@ -1,0 +1,257 @@
+package provider
+
+import (
+	"context"
+	"fmt"
+	"io/ioutil"
+	"strconv"
+
+	dd "github.com/doximity/defect-dojo-client-go"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-go/tftypes"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+)
+
+type productTypeResourceType struct{}
+
+func (t productTypeResourceType) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diagnostics) {
+	return tfsdk.Schema{
+		// This description is used by the documentation generator and the language server.
+		MarkdownDescription: "DefectDojo Product Type",
+
+		Attributes: map[string]tfsdk.Attribute{
+			"name": {
+				MarkdownDescription: "The name of the Product Type",
+				Required:            true,
+				Type:                types.StringType,
+			},
+			"description": {
+				MarkdownDescription: "The description of the Product Type",
+				Required:            true,
+				Type:                types.StringType,
+			},
+			"id": {
+				Computed:            true,
+				MarkdownDescription: "Identifier",
+				PlanModifiers: tfsdk.AttributePlanModifiers{
+					tfsdk.UseStateForUnknown(),
+				},
+				Type: types.StringType, // the id (for import purposes) MUST be a string
+			},
+		},
+	}, nil
+}
+
+func (t productTypeResourceType) NewResource(ctx context.Context, in tfsdk.Provider) (tfsdk.Resource, diag.Diagnostics) {
+	provider, diags := convertProviderType(in)
+
+	return productTypeResource{
+		provider: provider,
+	}, diags
+}
+
+type productTypeResourceData struct {
+	Name        types.String `tfsdk:"name"`
+	Description types.String `tfsdk:"description"`
+	Id          types.String `tfsdk:"id"`
+}
+
+type productTypeResource struct {
+	provider provider
+}
+
+func (r productTypeResource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse) {
+	var data productTypeResourceData
+
+	diags := req.Config.Get(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	apiResp, err := r.provider.client.ProductTypesCreateWithResponse(ctx, dd.ProductTypesCreateJSONRequestBody{
+		Description: &data.Description.Value,
+		Name:        data.Name.Value,
+	})
+
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error Creating Resource",
+			fmt.Sprintf("%s", err))
+		return
+	}
+
+	if apiResp.StatusCode() == 201 {
+		data.Id = types.String{Value: fmt.Sprint(apiResp.JSON201.Id)}
+		data.Name = types.String{Value: apiResp.JSON201.Name}
+		data.Description = types.String{Value: *apiResp.JSON201.Description}
+	} else {
+		body, _ := ioutil.ReadAll(apiResp.HTTPResponse.Body)
+
+		resp.Diagnostics.AddError(
+			"API Error Creating Resource",
+			fmt.Sprintf("Unexpected response code from API: %d", apiResp.StatusCode())+
+				fmt.Sprintf("\n\nbody:\n\n%s", body),
+		)
+	}
+
+	// write logs using the tflog package
+	// see https://pkg.go.dev/github.com/hashicorp/terraform-plugin-log/tflog
+	// for more information
+	tflog.Trace(ctx, "created a product type")
+
+	diags = resp.State.Set(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+}
+
+func (r productTypeResource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, resp *tfsdk.ReadResourceResponse) {
+	var data productTypeResourceData
+
+	diags := req.State.Get(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if data.Id.Null {
+		resp.Diagnostics.AddError(
+			"Could not Retrieve Resource",
+			"The Id field was null but it is required to retrieve the product type")
+	}
+
+	idNumber, err := strconv.Atoi(data.Id.Value)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Could not Retrieve Resource",
+			fmt.Sprintf("Error while parsing the Product Type ID from state: %s", err))
+	}
+
+	apiResp, err := r.provider.client.ProductTypesRetrieveWithResponse(ctx, idNumber, &dd.ProductTypesRetrieveParams{})
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error Retrieving Resource",
+			fmt.Sprintf("%s", err))
+		return
+	}
+
+	if apiResp.StatusCode() == 200 {
+		data.Name = types.String{Value: apiResp.JSON200.Name}
+		data.Description = types.String{Value: *apiResp.JSON200.Description}
+	} else {
+		body, _ := ioutil.ReadAll(apiResp.HTTPResponse.Body)
+
+		resp.Diagnostics.AddError(
+			"API Error Retrieving Resource",
+			fmt.Sprintf("Unexpected response code from API: %d", apiResp.StatusCode())+
+				fmt.Sprintf("\n\nbody:\n\n%+v", body),
+		)
+	}
+
+	diags = resp.State.Set(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+}
+
+func (r productTypeResource) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
+	var data productTypeResourceData
+
+	diags := req.Plan.Get(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if data.Id.Null {
+		resp.Diagnostics.AddError(
+			"Could not Update Resource",
+			"The Id field was null but it is required to retrieve the product")
+	}
+
+	idNumber, err := strconv.Atoi(data.Id.Value)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Could not Update Resource",
+			fmt.Sprintf("Error while parsing the Product Type ID from state: %s", err))
+	}
+
+	apiResp, err := r.provider.client.ProductTypesUpdateWithResponse(ctx, idNumber, dd.ProductTypesUpdateJSONRequestBody{
+		Description: &data.Description.Value,
+		Name:        data.Name.Value,
+	})
+
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error Updating Resource",
+			fmt.Sprintf("%s", err))
+		return
+	}
+
+	if apiResp.StatusCode() == 200 {
+		data.Id = types.String{Value: fmt.Sprint(apiResp.JSON200.Id)}
+		data.Name = types.String{Value: apiResp.JSON200.Name}
+		data.Description = types.String{Value: *apiResp.JSON200.Description}
+	} else {
+		body, _ := ioutil.ReadAll(apiResp.HTTPResponse.Body)
+
+		resp.Diagnostics.AddError(
+			"API Error Updating Resource",
+			fmt.Sprintf("Unexpected response code from API: %d", apiResp.StatusCode())+
+				fmt.Sprintf("\n\nbody:\n\n%+v", body),
+		)
+	}
+
+	diags = resp.State.Set(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+}
+
+func (r productTypeResource) Delete(ctx context.Context, req tfsdk.DeleteResourceRequest, resp *tfsdk.DeleteResourceResponse) {
+	var data productTypeResourceData
+
+	diags := req.State.Get(ctx, &data)
+	resp.Diagnostics.Append(diags...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if data.Id.Null {
+		resp.Diagnostics.AddError(
+			"Could not Delete Resource",
+			"The Id field was null but it is required to retrieve the product type")
+	}
+
+	idNumber, err := strconv.Atoi(data.Id.Value)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Could not Delete Resource",
+			fmt.Sprintf("Error while parsing the Product Type ID from state: %s", err))
+	}
+
+	apiResp, err := r.provider.client.ProductTypesDestroy(ctx, idNumber)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error Deleting Resource",
+			fmt.Sprintf("%s", err))
+		return
+	}
+
+	if apiResp.StatusCode != 204 {
+		body, _ := ioutil.ReadAll(apiResp.Body)
+
+		resp.Diagnostics.AddError(
+			"API Error Deleting Resource",
+			fmt.Sprintf("Unexpected response code from API: %d", apiResp.StatusCode)+
+				fmt.Sprintf("\n\nbody:\n\n%+v", body),
+		)
+	}
+
+	resp.State.RemoveResource(ctx)
+}
+
+func (r productTypeResource) ImportState(ctx context.Context, req tfsdk.ImportResourceStateRequest, resp *tfsdk.ImportResourceStateResponse) {
+	tfsdk.ResourceImportStatePassthroughID(ctx, tftypes.NewAttributePath().WithAttributeName("id"), req, resp)
+}