+++ internal/provider/provider.go
@@ -3,10 +3,14 @@ package provider
 import (
 	"context"
 	"fmt"
+	"os"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+
+	"github.com/deepmap/oapi-codegen/pkg/securityprovider"
+	dd "github.com/doximity/defect-dojo-client-go"
 )
 
 // provider satisfies the tfsdk.Provider interface and usually is included
@@ -15,9 +19,7 @@ type provider struct {
 	// client can contain the upstream provider SDK or HTTP client used to
 	// communicate with the upstream service. Resource and DataSource
 	// implementations can then make calls using this client.
-	//
-	// TODO: If appropriate, implement upstream provider SDK or HTTP client.
-	// client vendorsdk.ExampleClient
+	client *dd.ClientWithResponses
 
 	// configured is set to true at the end of the Configure method.
 	// This can be used in Resource and DataSource implementations to verify
@@ -32,7 +34,10 @@ type provider struct {
 
 // providerData can be used to store data from the Terraform configuration.
 type providerData struct {
-	Example types.String `tfsdk:"example"`
+	BaseUrl  types.String `tfsdk:"base_url"`
+	ApiKey   types.String `tfsdk:"api_key"`
+	Username types.String `tfsdk:"username"`
+	Password types.String `tfsdk:"password"`
 }
 
 func (p *provider) Configure(ctx context.Context, req tfsdk.ConfigureProviderRequest, resp *tfsdk.ConfigureProviderResponse) {
@@ -44,34 +49,155 @@ func (p *provider) Configure(ctx context.Context, req tfsdk.ConfigureProviderReq
 		return
 	}
 
-	// Configuration values are now available.
-	// if data.Example.Null { /* ... */ }
+	var (
+		url   string
+		token string
+		user  string
+		pass  string
+	)
 
-	// If the upstream provider SDK or HTTP client requires configuration, such
-	// as authentication or logging, this is a great opportunity to do so.
+	if data.BaseUrl.Null {
+		url = os.Getenv("DEFECTDOJO_BASEURL")
+	} else {
+		url = data.BaseUrl.Value
+	}
+
+	if url == "" {
+		resp.Diagnostics.AddError(
+			"Unable to configure provider",
+			"Could not determine the url of the defectdojo service. No base_url value provided and no DEFECTDOJO_BASEURL environment variable.",
+		)
+		return
+	}
+
+	if !data.ApiKey.Null {
+		token = data.ApiKey.Value
+	} else {
+		token = os.Getenv("DEFECTDOJO_APIKEY")
+	}
+
+	if !data.Username.Null {
+		user = data.Username.Value
+	} else {
+		user = os.Getenv("DEFECTDOJO_USERNAME")
+	}
+
+	if !data.Password.Null {
+		pass = data.Password.Value
+	} else {
+		pass = os.Getenv("DEFECTDOJO_PASSWORD")
+	}
+
+	// if we have no api key but we do have a user and password explicitly set,
+	// we can use the API to get the token dynamically instead.
+	if token == "" && user != "" && pass != "" {
+		tokenclient, err := dd.NewClientWithResponses(url)
+		if err != nil {
+			resp.Diagnostics.AddError(
+				"Unable to configure provider",
+				"Error instantiating the client. This should never happen.",
+			)
+			return
+		}
+
+		tokenResponse, err := tokenclient.ApiTokenAuthCreateWithResponse(ctx, dd.ApiTokenAuthCreateJSONRequestBody{
+			Username: user,
+			Password: pass,
+		})
+
+		if err != nil {
+			resp.Diagnostics.AddError(
+				"Unable to configure provider",
+				fmt.Sprintf("Network error retrieving the token via the API: %s", err),
+			)
+			return
+		}
+
+		if tokenResponse.StatusCode() == 200 {
+			token = tokenResponse.JSON200.Token
+		} else {
+			resp.Diagnostics.AddError(
+				"Unable to configure provider",
+				fmt.Sprintf("Error retrieving the api token via the API. Unexpected response code: %d", tokenResponse.StatusCode()),
+			)
+			return
+		}
+	}
+
+	if token == "" {
+		resp.Diagnostics.AddError(
+			"Unable to configure provider",
+			"Could not determine the api key for the defectdojo service. No api_key value provided and no DEFECTDOJO_APIKEY environment variable.",
+		)
+		return
+	}
+
+	tokenProvider, err := securityprovider.NewSecurityProviderApiKey("header", "Authorization", fmt.Sprintf("Token %s", token))
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Unable to configure provider",
+			"Error instantiating the security provider. This should never happen.",
+		)
+		return
+	}
+
+	client, err := dd.NewClientWithResponses(url, dd.WithRequestEditorFn(tokenProvider.Intercept))
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Unable to configure provider",
+			"Error instantiating the client. This should never happen.",
+		)
+		return
+	}
+
+	p.client = client
 
 	p.configured = true
 }
 
 func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {
 	return map[string]tfsdk.ResourceType{
-		"scaffolding_example": exampleResourceType{},
+		"defectdojo_product":      productResourceType{},
+		"defectdojo_product_type": productTypeResourceType{},
 	}, nil
 }
 
 func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSourceType, diag.Diagnostics) {
 	return map[string]tfsdk.DataSourceType{
-		"scaffolding_example": exampleDataSourceType{},
+		"defectdojo_product":      productDataSourceType{},
+		"defectdojo_product_type": productTypeDataSourceType{},
 	}, nil
 }
 
 func (p *provider) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diagnostics) {
 	return tfsdk.Schema{
 		Attributes: map[string]tfsdk.Attribute{
-			"example": {
-				MarkdownDescription: "Example provider attribute",
+			"base_url": {
+				MarkdownDescription: "Base URL of the defectdojo installation",
 				Optional:            true,
+				Required:            false,
 				Type:                types.StringType,
+			},
+			"api_key": {
+				MarkdownDescription: "The API Key used to authenticate to defectdojo",
+				Optional:            true,
+				Required:            false,
+				Type:                types.StringType,
+				Sensitive:           true,
+			},
+			"username": {
+				MarkdownDescription: "The username used to authenticate to defectdojo. Has no effect if api_key is set.",
+				Optional:            true,
+				Required:            false,
+				Type:                types.StringType,
+				Sensitive:           true,
+			},
+			"password": {
+				MarkdownDescription: "The password used to authenticate to defectdojo. Has no effect if api_key is set.",
+				Optional:            true,
+				Required:            false,
+				Type:                types.StringType,
+				Sensitive:           true,
 			},
 		},
 	}, nil