+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.1.0 (Unreleased)
+## 0.0.4
 
 FEATURES:
+ - Add basic support for Product Type resource and data source
+
+## 0.0.3
+
+FEATURES:
+ - First working version.
+ - Basic support for Product resource and data source.